+++ src/controladores/fastapi/usuario/c_remover_endereco_fastapi.py
