+++ tests/controladores/fastapi/usuario/test_c_adicionar_telefone_fastapi.py
@@ -1,7 +1,7 @@
 from devmaua.src.models.erros.erro_telefone import ErroDadosTelefoneInvalidos
 from devmaua.src.models.erros.erro_usuario import ErroDadosUsuarioInvalidos
 
-from src.controladores.fastapi.c_adicionar_telefone_fastapi import ControllerHTTPAdicionarTelefoneFastAPI
+from src.controladores.fastapi.usuario.c_adicionar_telefone_fastapi import ControllerHTTPAdicionarTelefoneFastAPI
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
 from src.usecases.erros.erros_uc_alteracao_info_cadastro_usuario import ErroUsuarioNaoExiste
 from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
