+++ tests/controladores/fastapi/usuario/test_c_deletar_usuario_por_email_fastapi.py
@@ -1,11 +1,13 @@
-from src.controladores.fastapi.c_editar_email_fastapi import ControllerHTTPEditarEmailFastAPI
+from src.controladores.fastapi.usuario.c_deletar_usuario_por_email_fastapi import CDeletarUsuarioPorEmailFastAPI
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
 from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
 
 from devmaua.src.models.usuario import Usuario
 
+from src.usecases.erros.erros_uc_alteracao_info_cadastro_usuario import ErroUsuarioNaoExiste
 
-class TestControllerEditarEmailFastAPI():
+
+class TestControllerDeletarUsuarioPorEmailFastAPI():
     
     def mockDictUsuario(self):
         return {
@@ -24,13 +26,7 @@ class TestControllerEditarEmailFastAPI():
                         "email": "teste@teste.com",
                         "tipo": 1,
                         "prioridade": 1
-                    },
-                    {
-                        "email": "novo@email.com",
-                        "tipo": 2,
-                        "prioridade": 2
                     }
-    
                 ],
                 "enderecos": [
                     {
@@ -48,13 +44,6 @@ class TestControllerEditarEmailFastAPI():
             ]
         }
         
-    def mockDictEmail(self):
-        return {
-                "email": "novo@email.com",
-                "tipo": 2,
-                "prioridade": 2
-                }
-    
     def mockRepositorioComUmUsuario(self) -> ArmazenamentoUsuarioVolatil:
         repositorio = ArmazenamentoUsuarioVolatil()
         cadastrador = UCCadastrarUsuario(repositorio)
@@ -64,19 +53,25 @@ class TestControllerEditarEmailFastAPI():
         return repositorio
     
             
-    def test_controller_editar_email_fastapi(self):
+    def test_controller_deletar_usuario_por_email_fastapi(self):
         repoVolatil = self.mockRepositorioComUmUsuario()
-        controllerEditarEmailFastAPI = ControllerHTTPEditarEmailFastAPI(repoVolatil)
+        controllerDeletarUsuarioPorEmail = CDeletarUsuarioPorEmailFastAPI(repoVolatil)
         
-        usuarioDict = self.mockDictUsuario()
-        emailDict = self.mockDictEmail()
         body = {
-                "usuario": usuarioDict,
-                "email": emailDict,
-                "emailNovo": "email@editado.br",
-                "tipo": 3,
-                "prioridade": 3
-                }
-        response = controllerEditarEmailFastAPI(body = body)
-
+                    "email": "teste@teste.com"
+               }
+        
+        response = controllerDeletarUsuarioPorEmail(body = body)
         assert response.status_code == 200
+        assert repoVolatil.usuarioExistePorEmail("teste@teste.com") == False
+        
+    def test_erro_usuario_inexistente(self):
+        repoVolatil = ArmazenamentoUsuarioVolatil()
+        controllerDeletarUsuarioPorEmail = CDeletarUsuarioPorEmailFastAPI(repoVolatil)
+        
+        body = {
+                    "email": "teste@teste.com"
+               }
+        response = controllerDeletarUsuarioPorEmail(body = body)
+        assert response.body.decode() == str(ErroUsuarioNaoExiste())
+        assert response.status_code == 404
