+++ tests/controladores/fastapi/usuario/test_c_editar_endereco_fastapi.py
@@ -1,11 +1,11 @@
-from src.controladores.fastapi.c_editar_telefone_fastapi import ControllerHTTPEditarTelefoneFastAPI
+from src.controladores.fastapi.usuario.c_editar_endereco_fastapi import ControllerHTTPEditarEnderecoFastAPI
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
 from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
 
 from devmaua.src.models.usuario import Usuario
 
 
-class TestControllerEditarTelefoneFastAPI():
+class TestControllerEditarEnderecoFastAPI():
     
     def mockDictUsuario(self):
         return {
@@ -43,13 +43,14 @@ class TestControllerEditarTelefoneFastAPI():
             ]
         }
         
-    def mockDictTelefone(self):
+    def mockDictEndereco(self):
         return {
-                    "tipo": 2,
-                    "numero": "99999-9999",
-                    "ddd": 11,
-                    "prioridade": 3
-                }
+                        "logradouro": "rua de tal",
+                        "numero": 20,
+                        "cep": "00000-000",
+                        "complemento": None,
+                        "tipo": 1
+                    }
     
     def mockRepositorioComUmUsuario(self) -> ArmazenamentoUsuarioVolatil:
         repositorio = ArmazenamentoUsuarioVolatil()
@@ -60,20 +61,21 @@ class TestControllerEditarTelefoneFastAPI():
         return repositorio
     
             
-    def test_controller_editar_telefone_fastapi(self):
+    def test_controller_editar_endereco_fastapi(self):
         repoVolatil = self.mockRepositorioComUmUsuario()
-        controllerEditarTelefoneFastAPI = ControllerHTTPEditarTelefoneFastAPI(repoVolatil)
+        controllerEditarEnderecoFastAPI = ControllerHTTPEditarEnderecoFastAPI(repoVolatil)
         
         usuarioDict = self.mockDictUsuario()
-        telefoneDict = self.mockDictTelefone()
+        enderecoDict = self.mockDictEndereco()
         body = {
                 "usuario": usuarioDict,
-                "telefone": telefoneDict,
-                "tipo": 3,
-                "ddd": 12,
-                "numero": '99999-8888',
-                "prioridade": 3
+                "endereco": enderecoDict,
+                "logradouro": "rua nova",
+                "numero": 201,
+                "cep": "00000-021",
+                "complemento": "Atras de outra rua",
+                "tipo": 2
             }
-        response = controllerEditarTelefoneFastAPI(body = body)
+        response = controllerEditarEnderecoFastAPI(body = body)
         
         assert response.status_code == 200
