+++ tests/controladores/fastapi/usuario/test_c_get_usuario_por_userid.py
@@ -1,7 +1,7 @@
 from devmaua.src.models.usuario import Usuario
 
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
-from src.controladores.fastapi.c_get_usuario_por_userid import CHttpGetUsuarioPorUserIdFastAPI
+from src.controladores.fastapi.usuario.c_get_usuario_por_userid import CHttpGetUsuarioPorUserIdFastAPI
 import tests.mock_objetos as mo
 import json
 import pytest
